test: split Realtime session tests by responsibility
