+++ .changeset/realtime-session-test-organization.md
@@ -1,0 +1,5 @@
+---
+'@openai/agents-realtime': patch
+---
+
+chore: organize Realtime session tool execution, approval, and timeout tests by responsibility.
